refactor(ui): colocate FilesPanel logic in controller hook
